ai_trigger/requirements.txt
Flask
gunicorn
google-cloud-run
requests